Fix issues with strict equality comparison
PHPStan has identified places where switching to strict equality
comparison has resulted in always-true or always-false scenarios. For
example, if we strictly compare a float to "0", it will always return
false because "0" is an int (whereas "0.0" is the appropriate float).

* In `get_colored_text_range`, require that the min/max values are
  ints instead of floats. We don't need this flexibility and it makes
  the equality comparison more robust.

* When comparing a numeric-string to an int in `smrBBCode` "race"
  condition, cast to an int before comparing to an int.

More issues with strict comparison may yet arise, but they will not be
discovered by static analysis.

+++ src/lib/Default/smr.inc.php
@@ -123,7 +123,7 @@ function smrBBCode(BBCode $bbParser, int $action, string $tagName, string $defau
 			case 'race':
 				$raceNameID = $default;
 				foreach (Race::getAllNames() as $raceID => $raceName) {
-					if ((is_numeric($raceNameID) && $raceNameID === $raceID)
+					if ((is_numeric($raceNameID) && (int)$raceNameID === $raceID)
 						|| $raceNameID === $raceName) {
 						if ($action === BBCode::BBCODE_CHECK) {
 							return true;
@@ -280,7 +280,7 @@ function create_submit_link(Page $container, string $text): string {
 	return '<a href="' . $container->href() . '" class="submitStyle">' . $text . '</a>';
 }
 
-function get_colored_text_range(float $value, float $maxValue, string $text = null, float $minValue = 0): string {
+function get_colored_text_range(float $value, int $maxValue, string $text = null, int $minValue = 0): string {
 	if ($text === null) {
 		$text = number_format($value);
 	}
